Anillo_Enero_2021_08_04_2021/requirements.txt
opcua==0.98.12
cassandra-driver==3.24.0
numpy==1.19.4
